Share RaceFlags
