+++ SharedSE/SharedSE.vcxitems
@@ -120,6 +120,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3AttachmentFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectType.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3RaceFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3SpellFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3WeaponFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)WindowsUtil.h" />
