+++ SharedSE/SharedSE.vcxitems.filters
@@ -357,7 +357,12 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3SpellFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
-    <ClInclude Include="$(MSBuildThisFileDirectory)TES3AttachmentFlags.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3AttachmentFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3RaceFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
